requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
openai==0.27.0
gtts==2.3.2
moviepy==1.0.3
Pillow==10.0.0
numpy==1.24.3
requests==2.32.4
gunicorn==21.2.0
python-dotenv==1.0.0 